+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -4,7 +4,7 @@ import io.ktor.server.application.Application
 import io.ktor.server.engine.connector
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
-import no.nav.syfo.application.api.apiModule
+import no.nav.syfo.application.api.configureRouting
 import no.nav.syfo.plugins.configureDependencies
 import no.nav.syfo.plugins.configureLifecycleHooks
 import org.koin.ktor.ext.get
@@ -35,6 +35,6 @@ fun main() {
 
 fun Application.module() {
     configureDependencies()
-    apiModule()
+    configureRouting()
     configureLifecycleHooks(get())
 }
