+++ src/main/kotlin/no/nav/syfo/plugins/Lifecycle.kt
@@ -4,9 +4,11 @@ import io.ktor.server.application.Application
 import io.ktor.server.application.ApplicationStarted
 import io.ktor.server.application.ApplicationStopped
 import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.dinesykmeldte.logger
+import no.nav.syfo.util.logger
 
 fun Application.configureLifecycleHooks(applicationState: ApplicationState = ApplicationState()) {
+    val logger = logger()
+
     monitor.subscribe(ApplicationStarted) {
         applicationState.ready = true
         logger.info("Application is ready, running Java VM ${Runtime.version()}")
